refactor:將註冊邏輯功能移動到父組件
